+++ src/lib/xml-parser.js
@@ -9,7 +9,8 @@
  * @returns {Article[]}
  */
 export function parseRssFeed (xmlStr, limit = 9) {
-  const doc = new DOMParser().parseFromString(xmlStr, 'application/xml');
+  const xmlStrTrimmed = xmlStr.trim();
+  const doc = new DOMParser().parseFromString(xmlStrTrimmed, 'application/xml');
   const error = doc.querySelector('parsererror');
   if (error != null) {
     throw new Error(error.innerHTML);